Use Double instead of Float in the portable ChatOptions
* The type of the portable chat options (frequencyPenalty, presencePenalty, temperature, topP) has been changed from Float to Double.
* APIs, test, and documentation have been updated accordingly.

Fixes gh-712

+++ models/spring-ai-anthropic/src/main/java/org/springframework/ai/anthropic/AnthropicChatModel.java
@@ -82,7 +82,7 @@ public class AnthropicChatModel extends AbstractToolCallSupport implements ChatM
 
 	public static final Integer DEFAULT_MAX_TOKENS = 500;
 
-	public static final Float DEFAULT_TEMPERATURE = 0.8f;
+	public static final Double DEFAULT_TEMPERATURE = 0.8;
 
 	/**
 	 * The lower-level API for the Anthropic service.
