+++ models/spring-ai-bedrock/src/main/java/org/springframework/ai/bedrock/anthropic/AnthropicChatOptions.java
@@ -39,7 +39,7 @@ public class AnthropicChatOptions implements ChatOptions {
 	 * responses from the generative. This value specifies default to be used by the backend while making the call to
 	 * the generative.
 	 */
-	private @JsonProperty("temperature") Float temperature;
+	private @JsonProperty("temperature") Double temperature;
 
 	/**
 	 * Specify the maximum number of tokens to use in the generated response. Note that the models may stop before
@@ -57,7 +57,7 @@ public class AnthropicChatOptions implements ChatOptions {
 	 * The maximum cumulative probability of tokens to consider when sampling. The generative uses combined Top-k and
 	 * nucleus sampling. Nucleus sampling considers the smallest set of tokens whose probability sum is at least topP.
 	 */
-	private @JsonProperty("top_p") Float topP;
+	private @JsonProperty("top_p") Double topP;
 
 	/**
 	 * Configure up to four sequences that the generative recognizes. After a stop sequence, the generative stops
@@ -79,7 +79,7 @@ public class AnthropicChatOptions implements ChatOptions {
 
 		private final AnthropicChatOptions options = new AnthropicChatOptions();
 
-		public Builder withTemperature(Float temperature) {
+		public Builder withTemperature(Double temperature) {
 			this.options.setTemperature(temperature);
 			return this;
 		}
@@ -94,7 +94,7 @@ public class AnthropicChatOptions implements ChatOptions {
 			return this;
 		}
 
-		public Builder withTopP(Float topP) {
+		public Builder withTopP(Double topP) {
 			this.options.setTopP(topP);
 			return this;
 		}
@@ -116,11 +116,11 @@ public class AnthropicChatOptions implements ChatOptions {
 	}
 
 	@Override
-	public Float getTemperature() {
+	public Double getTemperature() {
 		return this.temperature;
 	}
 
-	public void setTemperature(Float temperature) {
+	public void setTemperature(Double temperature) {
 		this.temperature = temperature;
 	}
 
@@ -153,11 +153,11 @@ public class AnthropicChatOptions implements ChatOptions {
 	}
 
 	@Override
-	public Float getTopP() {
+	public Double getTopP() {
 		return this.topP;
 	}
 
-	public void setTopP(Float topP) {
+	public void setTopP(Double topP) {
 		this.topP = topP;
 	}
 
@@ -186,13 +186,13 @@ public class AnthropicChatOptions implements ChatOptions {
 
 	@Override
 	@JsonIgnore
-	public Float getFrequencyPenalty() {
+	public Double getFrequencyPenalty() {
 		return null;
 	}
 
 	@Override
 	@JsonIgnore
-	public Float getPresencePenalty() {
+	public Double getPresencePenalty() {
 		return null;
 	}
